requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-filter==23.5
python-dotenv==1.0.0
Faker==22.0.0
